cloudformation/requirements.txt
-e .
aws-cdk.aws-iam==1.102.0
aws-cdk.aws-certificatemanager==1.102.0
aws-cdk.aws-cloudfront==1.102.0
aws-cdk.aws-route53==1.102.0
aws-cdk.aws-route53-targets==1.102.0
aws-cdk.aws-s3==1.102.0
aws-cdk.aws-s3-deployment==1.102.0
aws-cdk.aws-ssm==1.102.0
